[refactor] DTO를 도메인 별 패키지로 분리
