+++ core/data/src/main/java/com/ssavice/data/service/CompanyRetrofitService.kt
@@ -1,25 +1,24 @@
 package com.ssavice.data.service
 
 import GetCompanyInfoDTO
-import com.ssavice.network.model.AddCompanyDTO
-import com.ssavice.network.model.CompanyBookDTO
-import com.ssavice.network.model.CompanyBookSummaryDTO
-import com.ssavice.network.model.CompanyDetailDTO
+import com.ssavice.network.model.company.AddCompanyDTO
+import com.ssavice.network.model.company.CompanyBookDTO
+import com.ssavice.network.model.company.CompanyBookSummaryDTO
+import com.ssavice.network.model.company.CompanyDetailDTO
 import com.ssavice.network.model.ConfirmImageDTO
 import com.ssavice.network.model.ContentTypeDTO
 import com.ssavice.network.model.GetAddressDTO
-import com.ssavice.network.model.GetCompanyReviewDTO
-import com.ssavice.network.model.GetCompanySummaryDTO
+import com.ssavice.network.model.review.GetCompanyReviewDTO
+import com.ssavice.network.model.company.GetCompanySummaryDTO
 import com.ssavice.network.model.JwtDTO
 import com.ssavice.network.model.LoginDTO
 import com.ssavice.network.model.PresignedUrlDTO
-import com.ssavice.network.model.UpdateCompanyProfileDTO
-import com.ssavice.network.model.ValidateBusinessDTO
-import com.ssavice.network.model.ValidateBusinessResponseDTO
+import com.ssavice.network.model.company.UpdateCompanyProfileDTO
+import com.ssavice.network.model.company.ValidateBusinessDTO
+import com.ssavice.network.model.company.ValidateBusinessResponseDTO
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
-import retrofit2.http.Headers
 import retrofit2.http.POST
 import retrofit2.http.PUT
 import retrofit2.http.Path
