+++ core/data/src/main/java/com/ssavice/data/service/UserInfoRetrofitService.kt
@@ -5,13 +5,13 @@ import com.ssavice.network.model.ContentTypeDTO
 import com.ssavice.network.model.GetAddressDTO
 import com.ssavice.network.model.PresignedUrlDTO
 import com.ssavice.network.model.RegionPostDTO
-import com.ssavice.network.model.UpdateUserProfileDTO
-import com.ssavice.network.model.UpdateUserProfileResponseDTO
-import com.ssavice.network.model.UserBookDTO
-import com.ssavice.network.model.UserBookSummaryDTO
-import com.ssavice.network.model.UserProfileDTO
-import com.ssavice.network.model.WishListDTO
-import com.ssavice.network.model.WishServiceDTO
+import com.ssavice.network.model.user.UpdateUserProfileDTO
+import com.ssavice.network.model.user.UpdateUserProfileResponseDTO
+import com.ssavice.network.model.service.UserBookDTO
+import com.ssavice.network.model.user.UserBookSummaryDTO
+import com.ssavice.network.model.user.UserProfileDTO
+import com.ssavice.network.model.user.WishListDTO
+import com.ssavice.network.model.service.WishServiceDTO
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
