+++ core/network/src/main/kotlin/com/ssavice/network/model/company/GetCompanyInfoDTO.kt
