+++ core/network/src/main/kotlin/com/ssavice/network/model/review/PostReviewDTO.kt
@@ -1,4 +1,4 @@
-package com.ssavice.network.model
+package com.ssavice.network.model.review
 
 import com.ssavice.model.service.ReviewForm
 import kotlinx.serialization.Serializable
